Expose Traverson default message converters in a convenience method
The existing encapsulation of RestTemplate (via #201, #203) makes it the
user's responibility to set up the correct message converters (which as
argued in #203, is necessary), but it doesn't give any help knowing what
the converters should be.

This change exposes the default message converters via a static
convenience method, e.g.

    RestTemplate restTemplate = new RestTemplate();
    restTemplate.setInterceptors(...);
    restTemplate.setMessageConverters(
        Traverson.getDefaultMessageConverters(MediaTypes.HAL_JSON)
    );