#346 - Remove UriTemplate because it's unnecessary
Because the expansion happens at the bottom of getAndFindLinkWithRel, there is no need to use a UriTemplate. The first GET is the base URI, which isn't a template. The second hop will be a rel, and the expansion of that link happens (if it happens) at the bottom of getAndFindLinkWithRel, so when invoked in the next call, it won't be a template.