#560 - Add extra attributes to Link and use for HAL mediatype
Adds many additional attributes defined in RFC5988 and verifies they work properly in the neutral representation of Link while also being rendered properly in the HAL module.

Additionally, some of the files had inconsistent EOL characters. This patches that as well. To see the differences WITHOUT this bit, add ?w=1 to the github page showing the delta, and whitespace differences will be filtered out.

Resolves #100,#417,#235,#240,#238,#223