#828 - Polishing.
Switched to javadoc:javadoc goal as we're not an aggregator project and can prevent the lifecycle restart from being triggered which caused plugins being bound to lifecycle phases before compile to be re-triggered.

Upgraded to Asciidoctor Maven Plugin 1.5.7.1. Removed custom CSS stylesheet for Javadoc as it doesn't look right with current Javadoc.