#1304 - Properly handle custom mediatypes loaded from spring.factories.
If a project-provided mediatype should only be activated when explicitly requested via @EnableHypermediaSupport. However, a custom mediatype should never be filted out because of this.

To handle this, a marker interface is defined that all the project-specific mediatypes can implement. This was, the enabling code can test whether or not the project-specific mediatypes "support" the declared mediatypes. However, mediatypes that do NOT have this marker interface (custom ones), should be registered unconditionally.