#1435 - EntityModel now explicitly rejects types rendered as @JsonValue.
It doesn't make sense to wrap an object to be rendered as value into an EntityModel as the latter will end up as JSON Object and thus, the representation of the target object *needs* to consist of key-value pairs.

Previously we just produced invalid JSON which ultimately failed as well.