#1447 - Overhaul Jackson customizations for HAL FORMS.
Got rid of quite a few custom serializers in Jackson2HalFormsModule. We now rely on the HAL setup mostly and register a virtual bean property to render the _templates field. The Jackson component used for that is registered as Spring bean so that we can now use the HAL HandlerInstantiator with the Jackson2HalFormsModule.

Moved quite a few integration tests to the COntextualMapper API.