#1515 - Fix HAL link serialization issues when Jackson map entry sorting is enabled.
Prior to this commit, the map of links to be rendered would key by LinkRelation. Unfortunately, Jackson requires `Map` keys to implement `Comparable` in case `SerializationFeature.ORDER_MAP_ENTRIES_BY_KEYS` is enabled. We now use the link relation's value as key right aways as Strings are comparable out of the box.

Switched to use Spring's MultiValueMap to avoid having to deal with the value list initialization ourselves.